MatrixPilot_trunk: added correct SILSIM support for INITIALIZE_VERTICAL for VTOL option. corrected include in flighplan.c to restore build.
git-svn-id: https://gentlenav.googlecode.com/svn/trunk@3450 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ libDCM/estAltitude.c
@@ -57,7 +57,7 @@ void altimeter_calibrate(void)
 }
 
 #if (BAROMETER_ALTITUDE == 1)
-void udb_barometer_callback(long pressure, int temperature, char status)
+void udb_barometer_callback(long pressure, int16_t temperature, char status)
 {
 	barometer_temperature = temperature; // units of 0.1 deg C
 	barometer_pressure = pressure; // units are Pascals so this could be reduced to an uint16_t
